Some revisions for your file
I haven't seen your CSS file yet but so far you've got the idea. Some things you are doing right is the organization (A lot of your code is nested properly), the use of your API's, and the comments you're making. Some things that can be improved on is the syntax of HTML, the use of comments tags, and the obvious lack of code. Even though there isn't much here, you're doing awesome. KEEP GOING! YOU'VE GOT THIS!